refactor: 불필요한 useclient, 브릿지 정리, src아래로 컴포넌트 폴더 이동 (#185)
