--- components/common/Header/HeaderClient.tsx
@@ -1,7 +1,0 @@
-"use client";
-
-import Header, { HeaderProps } from "./Header";
-
-export default function HeaderClient(props: HeaderProps) {
-  return <Header {...props} />;
-}
