--- components/common/SearchInput/SearchInputClient.tsx
@@ -1,7 +1,0 @@
-"use client"
-
-import SearchInputClient from "./SearchInputClient";
-
-export default function SearchInput() {
-  return <SearchInputClient />;
-}
