+++ src/components/common/ModalPopup/Modal.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import Button from "../Button/Button";
 
 import CheckIcon from "@/src/assets/check.svg";
