+++ src/components/common/StoreImage/StoreImage.md
