+++ src/components/common/StoreImage/StoreImage.tsx
