+++ src/types/svg.d.ts
